Print basic index information, including the tree extension (#293)
The latter is the one we have to maintain/create when creating
and index from a tree.
